requirements.txt
# FastAPI and ASGI server
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.20
alembic>=1.12.0

# Vector Database
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Slack
slack-sdk>=3.21.3
slack-bolt>=1.18.0

# Claude/Anthropic
anthropic>=0.5.0

# Utilities
python-multipart>=0.0.6
redis>=4.6.0
requests>=2.31.0
aiohttp>=3.8.5

# Google Drive integration
google-api-python-client>=2.99.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Process management for MCP integration
psutil>=5.9.5

# Date handling
python-dateutil>=2.8.2

# PDF processing
pypdf>=3.16.2
markdown>=3.4.4